+++ src/drivers/beanstalk/Command.php
@@ -51,7 +51,7 @@ class Command extends CliCommand
      */
     public function actionRun()
     {
-        $this->queue->run();
+        $this->queue->run(false);
     }
 
     /**
@@ -60,7 +60,7 @@ class Command extends CliCommand
      */
     public function actionListen()
     {
-        $this->queue->listen();
+        $this->queue->run(true, 3);
     }
 
     /**