+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Yii2 Queue Extension Change Log
 - Bug #92: Resolve issue in debug panel (farmani-eigital)
 - Enh: Start and stop events of a worker (zhuravljov)
 - Bug #99: Retry connecting after connection has timed out for redis driver (cebe)
+- Enh #185: Loop object instead of Signal helper (zhuravljov)
 
 ## 2.0.1, November 13, 2017